% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{as.CELL}
\alias{as.CELL}
\alias{as.CELL,matrix-method}
\title{Coercion to a CELL object}
\usage{
as.CELL(object, name = "cell")

\S4method{as.CELL}{matrix}(object, name = "cell")
}
\arguments{
\item{object}{a numeric matrix}

\item{name}{a character specifying the internal name of the CELL object to create}
}
\value{
a S4 object of class CELL
}
\description{
Coerces a numeric matrix into a CELL object.

This function transforms a numeric matrix into one or several cell profiles.
}
\details{
The matrix must have its column names corresponding to the cell markers.
}

