% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{as.CLUSTER}
\alias{as.CLUSTER}
\alias{as.CLUSTER,CELL-method}
\alias{as.CLUSTER,matrix-method}
\title{Coercion to a CLUSTER object}
\usage{
as.CLUSTER(object, name = object@name, cluster = NULL, bin.width = 0.05)

\S4method{as.CLUSTER}{CELL}(object, name = object@name, cluster = NULL,
  bin.width = 0.05)

\S4method{as.CLUSTER}{matrix}(object, name = "cell_cluster", cluster = NULL,
  bin.width = 0.05)
}
\arguments{
\item{object}{a CELL object or a numeric matrix}

\item{name}{a character specifying the internal name of the CLUSTER object to create}

\item{cluster}{a character indicating a channel name that can be used to gather the cell profiles into several cluster profiles. If a channel named is specified then the created CLUSTER object will contain as many profiles as different values present in this channel. If this parameter is NULL then a CLUSTER object with only one profile will be created}

\item{bin.width}{a numeric value indicating the width of the bins in the density estimation computations (default=0.05)}
}
\value{
a S4 object of class CLUSTER
}
\description{
Coerces a CELL object or a numeric matrix into a CLUSTER object.

This function transforms the cell profiles from a CELL object or from a numeric matrix into one or several cell cluster profiles by computing the means, the standard deviations, and the densities of each marker.
}
\details{
The 'cluster' parameter is especially useful when importing FCS files containing the SPADE clustering results (where an additional channel is used to indicate the associations between cells and cell clusters) or FCS files from any other automatic gating algorithm.

In the context of a numeric matrix coercion, the matrix must have its column names corresponding to the cell markers.
}
