% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{as.GATE}
\alias{as.GATE}
\alias{as.GATE,CELL-method}
\alias{as.GATE,CLUSTER-method}
\title{Coercion to a GATE object}
\usage{
as.GATE(object, name = object@name, quantiles = c(0.01, 0.99))

\S4method{as.GATE}{CELL}(object, name = object@name, quantiles = c(0.01,
  0.99))

\S4method{as.GATE}{CLUSTER}(object, name = object@name, quantiles = c(0.01,
  0.99))
}
\arguments{
\item{object}{a CELL or CLUSTER object}

\item{name}{a character specifying the internal name of the GATE object to create}

\item{quantiles}{a numeric vector of two values specifying the quantiles to use for the computations of marker expression ranges}
}
\value{
a S4 object of class GATE
}
\description{
Coerces a CELL or CLUSTER object into a GATE object.

This function transforms cell or cell cluster profiles from a CELL or CLUSTER object into one or several gate profiles by computing the range of each marker.
}
\details{
By default the expression ranges are computed based on the 0.01 and 0.99 quantiles of the marker expression densities, but can be specified by the user. If quantiles is set to 0 and 1, then the expression ranges will correspond to the minimal and maximal values of the expression markers.

If several cell cluster profiles are present in a CLUSTER object, then the resulting GATE object will contain as many gate profiles.
}

