% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biplot}
\alias{biplot}
\title{Biplot representation of a CELL object}
\usage{
biplot(cell, marker1, marker2, default.min = -3, return.gg = FALSE)
}
\arguments{
\item{cell}{a CELL object}

\item{marker1}{a character indicating the marker name of the first dimension}

\item{marker2}{a character indicating the marker name of the second dimension}

\item{default.min}{a numeric value indicating the lower bound of the biplot representation}

\item{return.gg}{a logical indicating if the function should return a list of ggplot objects}
}
\value{
if return.gg is TRUE, the function returns a list of ggplot objects
}
\description{
Generates a biplot representation for the cell profiles stored in a CELL object.
}
\details{
In such representation, each dot corresponds to a cell profile and dots are plotted in a 2-dimensional space corresponding to the marker expressions.
}

