% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{c}
\alias{c}
\alias{c,CELL-method}
\alias{c,CLUSTER-method}
\alias{c,RES-method}
\title{Combination of CytoCompare objects}
\usage{
\S4method{c}{CELL}(x, ..., recursive = FALSE)

\S4method{c}{CLUSTER}(x, ..., recursive = FALSE)

\S4method{c}{RES}(x, ..., recursive = FALSE)
}
\arguments{
\item{x}{a first CELL, CLUSTER, GATE or RES object}

\item{...}{further objects of the same class as x to be combined}

\item{recursive}{a logical value indicating if the function recursively descends through lists combining all their elements into a vector. Not implemented and should be set to FALSE}
}
\value{
a S4 object of class CELL, CLUSTER, GATE or RES
}
\description{
Combines two or several CELL, CLUSTER, GATE or RES objects.
}
\details{
All the different objects to combine must be of the same type.

This function is especially useful when combining comparison results from different RES objects into a single RES object (the RES objects must share the same markers and marker weights). RES objects can be combined to an empty RES object (i.e. RES()).

This function is also especially useful when combining cell profiles obtained from different FCS files into one single CELL object.
}

