% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdf.density}
\alias{cdf.density}
\title{Cumulative distribution function of a DENSITY object}
\usage{
cdf.density(density, values)
}
\arguments{
\item{density}{a DENSITY object}

\item{values}{a numeric value or a numeric vector specifying the densities to compute}
}
\value{
a numeric value of the cumulative distribution values
}
\description{
Provides the cumulative distribution function (CDF) of a DENSITY object at specific values.
}
\details{
This function is used internally when comparing the marker expression densities of cluster profiles with the default comparison approach. This function can also be used by users willing to define their own statistical functions when comparing cell, cell cluster or gate profiles.
}

