% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cdf.uniform}
\alias{cdf.uniform}
\title{Cumulative distribution function of a uniform distribution}
\usage{
cdf.uniform(bounds, value)
}
\arguments{
\item{bounds}{a numeric vector indicating the support (lower and upper bounds) of the uniform distribution}

\item{value}{a numeric value specifying the densities to compute}
}
\value{
a numeric value of the cumulative distribution value
}
\description{
Provides the cumulative distribution function (CDF) of a uniform distribution at a specific value.
}
\details{
This function is used internally when comparing the marker expression densities of gate profiles with the default comparison approach. This function can also be used by users willing to define their own statistical functions for comparing cell, cell cluster or gate profiles.
}

