% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CLUSTER-class}
\alias{CLUSTER-class}
\alias{CLUSTER}
\title{CLUSTER class definition}
\description{
CLUSTER is a S4 object containing one or several cell cluster profiles.
}
\details{
This object mainly stores for each cell cluster profile, the means, the standard deviations and the densities of each marker.
}
\section{Slots}{

\describe{
\item{\code{name}}{a character indicating the internal name of the CLUSTER object}

\item{\code{profiles}}{a character vector containing the names of the cell cluster profiles}

\item{\code{profiles.nb}}{an integer value indicating the number of cell cluster profiles}

\item{\code{profiles.sizes}}{an integer vector indicating the number of cells associated to each cluster profile}

\item{\code{markers}}{a character vector containing the marker names}

\item{\code{markers.nb}}{an integer value indicating the number of markers}

\item{\code{markers.clustering}}{a logical vector specifying the makers used as clustering markers}

\item{\code{means}}{a numeric matrix containing the means of each maker for each cluster profile}

\item{\code{sd}}{a numeric matrix containing the standard deviations of each maker for each cluster profile}

\item{\code{densities}}{a matrix of DENSITY objects containing the densities of each marker for each cluster profile}

\item{\code{overview.function}}{a character specifying the name of a function to call when plotting the CLUSTER object overview (please refer to the documentation of the 'plot()' function)}

\item{\code{graph}}{an object that can be used to store a visual representation of the cell clusters (e.g. a SPADE tree)}

\item{\code{graph.layout}}{a numeric matrix that can be used to store the positions of cell clusters in a 2-dimensional space (e.g. a SPADE tree layout)}
}}

