% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create.MWEIGHTS}
\alias{create.MWEIGHTS}
\title{Creation of a MWEIGHTS object}
\usage{
create.MWEIGHTS(markers)
}
\arguments{
\item{markers}{a character vector specifying the names of the markers}
}
\value{
a MWEIGHTS object containing the marker weights
}
\description{
Creates a MWEIGHTS object based on a set of marker names, where all marker weights are set to 1.
}
\details{
This function is a short-cut to the following code:\cr
markers        <- c("marker1","marker2","marker3","markeri...","markern")\cr
weights        <- rep(1,length(markers))\cr
mweights       <- MWEIGHTS(markers=markers,weights=weights)\cr\cr
The weight of each marker is set to 1 but can be changed afterwards using the function set.
}
