% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{DENSITY-class}
\alias{DENSITY-class}
\alias{DENSITY}
\title{DENSITY class definition}
\description{
DENSITY is a S4 object used to stores a marker expression density.
}
\details{
This object mainly stores for each marker: the bin characteristics, the negative and positive marker densities values, and the number of cells used in the density estimation. Densities are stored using two numeric vectors: 'values.neg' for the negative densities and 'values.pos' for the positive densities. This strategy allows to compute and store densities without defining an absolute minimal value or an absolute maximal value.
}
\section{Slots}{

\describe{
\item{\code{name}}{a character indicating the internal name of the CELL object}

\item{\code{bin.interval}}{a numeric vector of two values specifying the density boundaries}

\item{\code{bin.nb}}{a numeric vector of two values specifying the numbers of negative and positive bins}

\item{\code{values.pos}}{a numeric vector containing the positive density bins}

\item{\code{values.neg}}{a numeric vector containing the negative density bins}

\item{\code{point.nb}}{a numeric value indicating the number of point used to compute the expression density}

\item{\code{bin.width}}{a numeric value indicating the width of the bins used in the density estimation}
}}

