% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dheatmap}
\alias{dheatmap}
\title{Density heatmap of the marker expression}
\usage{
dheatmap(cluster, density.max = NULL, return.gg = FALSE)
}
\arguments{
\item{cluster}{a CLUSTER object containing one or several cell cluster profiles}

\item{density.max}{a numeric specifying the maximal density gradient value in the heatmap}

\item{return.gg}{a logical indicating if the function should return a list of ggplot objects}
}
\value{
if return.gg is TRUE, the function returns a list of ggplot objects
}
\description{
Generates a marker density heatmap for the cell cluster profiles stored in a `CLUSTER` object.
}
\details{
In such representation, each bar corresponds to a marker and the color gradient is proportional to the marker expression density.
}

