% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,CELL-method}
\title{Exportation of CELL objects}
\usage{
export(object, filename)

\S4method{export}{CELL}(object, filename = "cells.FCS")
}
\arguments{
\item{object}{a CELL object}

\item{filename}{a character indicating the location of the FCS to save}
}
\value{
none
}
\description{
Exports a CELL object into a FCS file.
}
