% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{GATE-class}
\alias{GATE}
\alias{GATE-class}
\title{GATE class definition}
\description{
GATE is a S4 object containing one or several gate profiles.
}
\details{
This object mainly stores for each gate profile, the intensity ranges of each marker.
}
\section{Slots}{

\describe{
\item{\code{name}}{a character indicating the internal name of the GATE object}

\item{\code{profiles}}{a character vector containing the names of the gate profiles}

\item{\code{profiles.nb}}{an integer value indicating the number of cell gate profiles}

\item{\code{markers}}{a character vector containing the marker names}

\item{\code{markers.nb}}{an integer value indicating the number of markers}

\item{\code{ranges}}{a 3-dimensional numeric array containing the intensity ranges of each marker for each gate profile}
}}

