% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.CELL}
\alias{import.CELL}
\title{Importation of cell profiles from a tab separated file}
\usage{
import.CELL(file, dictionary = NULL, exclude = NULL)
}
\arguments{
\item{file}{a character indicating the location of a tab separated file to import}

\item{dictionary}{a two-column data.frame providing the correspondence between the original marker names (first column) and the new marker names (second column)}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}
}
\value{
a S4 object of class CELL
}
\description{
Imports one or several cell profiles from a tab separated file into a CELL object.
}
\details{
Tab separated file to import must contain for each cell profile the intensities of the marker and must be formatted as the following:\cr
* each row must represent a cell profile;\cr
* each column must represent a marker;\cr
* each cell in the table must contain the marker expression intensities for a given cell profile;\cr
The first column must contain the cell names and the first row must contain the marker names.
}

