% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.FCS}
\alias{import.FCS}
\title{Importation of cell profiles from one or several FCS files}
\usage{
import.FCS(path, dictionary = NULL, exclude = NULL, trans = "arcsinh",
  trans.para = list(arcsinh.scale = 5), trans.exclude = NULL)
}
\arguments{
\item{path}{a character vector indicating the location to a FCS file or to a set of FCS files}

\item{dictionary}{a two-column data.frame providing the correspondence between the original marker names (first column) and the new marker names (second column)}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}

\item{trans}{a character specifying the name of a transformation function to apply on the marker expression intensities. Possible functions are "arcsinh" for arc sin hyperbolic transformation (default), "log" for logarithmic transformation, or "none" for no transformation}

\item{trans.para}{a character named list containing parameters for the transformation. Please refer to the details section for more details}

\item{trans.exclude}{a character vector containing the marker names for which no transformation will be applied on}
}
\value{
a S4 object of class CELL or CLUSTER
}
\description{
Imports one or several cell profiles from a FCS file or from a set of FCS files into a CELL object.
}
\details{
If a set of files is specified, then the files are merged in the import procedure.

Several transformations can be applied on the expression marker intensities via the 'trans' parameter:

The transformation functions can be parametrized using the named list 'trans.para'. The scale (cofactor) of the arcsinh transformation function can be parametrized using the 'arcsinh.scale' value. The shift of the log transformation function can be parametrized using the 'log.shift' value and the base of the log transformation function can be parametrized using the 'log.base' value.
}

