% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.GATE}
\alias{import.GATE}
\title{Importation of range gate profiles from a tab separated file}
\usage{
import.GATE(file, exclude = NULL)
}
\arguments{
\item{file}{a character indicating the location of a tab separated file to import}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}
}
\value{
a S4 object of class GATE
}
\description{
Imports one or several range gate profiles from a tab separated file into a GATE object.
}
\details{
Tab separated file to import must contain for each gate profile the ranges of the expression markers and must be formatted as the following:\cr
* each row must represent a gate profile;\cr
* each column must represent a marker;\cr
* each cell in the table must contain the marker expression lower and upper bounds for a given gate profile separated by a semicolon.\cr
The first column must contain the gate names and the first row must contain the marker names.
}

