% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.GATINGML}
\alias{import.GATINGML}
\title{Importation of  range gate profiles from a Gating-ML file}
\usage{
import.GATINGML(file, filterId = NULL)
}
\arguments{
\item{file}{a character indicating the location of a Gating-ML xml file to import}

\item{filterId}{a character vector indicating the identifiers of the gates to import}
}
\value{
a S4 object of class GATE
}
\description{
Imports one or several range gate profiles from a Gating-ML file into a GATE object.
}
\details{
Gating-ML is a standard file format for gate definitions developed to facilitate the interchange between different analysis software. Gating-ML rectangular gates, that are defined by ranges of expression markers, can be imported in CytoCompare using this function.\cr
}

