% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.VISNE}
\alias{import.VISNE}
\title{Importation of cell profiles from viSNE FCS files}
\usage{
import.VISNE(path, dictionary = NULL, exclude = NULL, trans = "arcsinh",
  trans.para = list(arcsinh.scale = 5), trans.exclude = NULL,
  tSNE1 = "tSNE1", tSNE2 = "tSNE2")
}
\arguments{
\item{path}{a character vector indicating the location to a viSNE FCS file or to a set of viSNE FCS files}

\item{dictionary}{a two-column data.frame providing the correspondence between the original marker names (first column) and the new marker names (second column)}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}

\item{trans}{a character specifying the name of a transformation function to apply on the marker expression intensities. Possible functions are "arcsinh" for arc sin hyperbolic transformation (default), "log" for logarithmic transformation, or "none" for no transformation}

\item{trans.para}{a character named list containing parameters for the transformation. Please refer to the details section for more details}

\item{trans.exclude}{a character vector containing the marker names for which no transformation will be applied on}

\item{tSNE1}{a character indicating the marker name of the first viSNE dimension (tSNE1)}

\item{tSNE2}{a character indicating the marker name of the second viSNE dimension (tSNE2)}
}
\value{
a S4 object of class CELL
}
\description{
Imports viSNE FCS files containing one or several cell profiles into a CELL object.
}
\details{
ViSNE is a dimensionality reduction algorithm designed for analysis and visualization of high-dimensionality cytometry data (PMID:23685480). In a viSNE map, each dot of the representation corresponds to a cell profile in a two-dimensional space (tSNE1 and tSNE2 dimensions).
}

