% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.VISNE_ACCENSE}
\alias{import.VISNE_ACCENSE}
\title{Importation of cell cluster profiles from viSNE/ACCENSE results}
\usage{
import.VISNE_ACCENSE(file, exclude = NULL, trans = "arcsinh",
  trans.para = switch(trans, arcsinh = list(arcsinh.scale = 5), log =
  list(log.shift = "auto", log.base = 10), none = NULL),
  trans.exclude = "population", bin.width = 0.05)
}
\arguments{
\item{file}{a character indicating the location to a zip or a folder containing the SPADE results}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}

\item{trans}{a character specifying the name of a transformation function to apply on the marker expression intensities. Possible functions are "arcsinh" for arc sin hyperbolic transformation (default), "log" for logarithmic transformation, or "none" for no transformation}

\item{trans.para}{a named list containing parameters for the transformation. Please refer to the details section for more details}

\item{trans.exclude}{a character vector containing the marker names for which no transformation must be applied on}

\item{bin.width}{a numeric value indicating the width of the bins for the marker expression densities computations}
}
\value{
a S4 object of class CLUSTER
}
\description{
Imports one or several cell cluster profiles identified by the viSNE/ACCENSE algorithm into a CLUSTER object.
}

