% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install.requiredpackages}
\alias{install.requiredpackages}
\title{Installation of the packages required by CytoCompare}
\usage{
install.requiredpackages()
}
\value{
none
}
\description{
Downloads and installs the R packages, from the CRAN (the Comprehensive R Archive Network) or Bioconductor, required to run CytoCompare.
}
\details{
This function install the following packages: diptest, ggplot2, ggrepel, grid, igraph, MASS, RJSONIO, XML, flowCore and flowUtils.
}

