% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{intersect}
\alias{intersect}
\alias{intersect,CELL,CELL-method}
\alias{intersect,CELL,CLUSTER-method}
\alias{intersect,CELL,GATE-method}
\alias{intersect,CLUSTER,CELL-method}
\alias{intersect,CLUSTER,CLUSTER-method}
\alias{intersect,CLUSTER,GATE-method}
\alias{intersect,GATE,CELL-method}
\alias{intersect,GATE,CLUSTER-method}
\alias{intersect,GATE,GATE-method}
\title{Identification of common markers between two cytometry objects}
\usage{
\S4method{intersect}{CELL,CELL}(x, y)

\S4method{intersect}{CLUSTER,CLUSTER}(x, y)

\S4method{intersect}{GATE,GATE}(x, y)

\S4method{intersect}{CELL,CLUSTER}(x, y)

\S4method{intersect}{CELL,GATE}(x, y)

\S4method{intersect}{CLUSTER,GATE}(x, y)

\S4method{intersect}{CLUSTER,CELL}(x, y)

\S4method{intersect}{GATE,CELL}(x, y)

\S4method{intersect}{GATE,CLUSTER}(x, y)
}
\arguments{
\item{x}{a CELL, CLUSTER or GATE object}

\item{y}{a CELL, CLUSTER or GATE object}
}
\value{
a S4 object of class MWEIGHTS
}
\description{
Identifies the common markers between two cytometry objects (CELL, CLUSTER or GATE objects) and store the results in a MWEIGHTS object.
}
\details{
The weight of each marker is set to 1 but can be changed afterwards using the function set.
}

