% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{MWEIGHTS-class}
\alias{MWEIGHTS-class}
\alias{MWEIGHTS}
\title{MWEIGHTS class definition}
\description{
MWEIGHTS is a S4 object containing the marker weights to use in the comparison computations.
}
\details{
This object mainly stores for each marker: the markers names and marker weights.
}
\section{Slots}{

\describe{
\item{\code{markers}}{a character vector containing the marker names}

\item{\code{weights}}{a numeric vector containing the marker weights}
}}

