% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,CLUSTER,CLUSTER-method}
\alias{plot,CLUSTER,missing-method}
\alias{plot,DENSITY,missing-method}
\alias{plot,MWEIGHTS,missing-method}
\alias{plot,RES,missing-method}
\title{Plot for all S4 CytoCompare objects}
\usage{
plot(object1, object2, ...)

\S4method{plot}{CLUSTER,missing}(object1, object2, overview = FALSE,
  return.gg = FALSE)

\S4method{plot}{CLUSTER,CLUSTER}(object1, object2, return.gg = FALSE)

\S4method{plot}{MWEIGHTS,missing}(object1, object2, return.gg = FALSE)

\S4method{plot}{DENSITY,missing}(object1, object2, return.gg = FALSE)

\S4method{plot}{RES,missing}(object1, return.gg = FALSE, ...)
}
\arguments{
\item{object1}{a CELL, CLUSTER, GATE, MWEIGHTS or DENSITY object to plot}

\item{object2}{another object to plot over the first object (only for CELL, CLUSTER or GATE objects)}

\item{...}{other parameters}

\item{overview}{a logical indicating is an overview of the object must be plotted (only available for single CELL and CLUSTER)}

\item{return.gg}{a logical indicating if the function should return a list of ggplot objects}
}
\value{
if return.gg is TRUE, the function returns a list of ggplot objects
}
\description{
Makes visual representations for CELL, CLUSTER, GATE, MWEIGHTS or DENSITY objects.
}
\details{
Profiles contained in CELL, CLUSTER, GATE objects can be represented alone (object2=NULL) or in combination with another CELL, CLUSTER, GATE objects (via object2).

Cell profiles are represented via parallel coordinates where the x-axis represents the different markers and where the y-axis represents the marker expressions.

Cell cluster profiles are represented via parallel coordinates where the x-axis represents the different markers, where the y-axis represents the marker expressions, and where error bars indicate the marker expression standard deviations.

Gate profiles are represented via ribbons where the x-axis represents the different markers and where the y-axis represents the marker intensity ranges

In the case of a single CELL object, the parameter 'overview' indicates if an overview of the CELL object must be represented (e.g. viSNE map). In the case of a single CLUSTER object, the parameter 'overview' indicates if an overview of the CLUSTER object must be represented (e.g. a SPADE tree). In both cases, the plot function will call the function indicated in the 'overview.function' slot of the CELL or CLUSTER objects. 

Marker weights contained in MWEIGHTS objects can be represented via bar plots where each bar corresponds to a marker and where the bar heights are proportional to the marker weights. 

Density profiles contained in DENSITY objects can be represented via histogram plots where each bar corresponds to a density bin and where a smooth line represents the average estimation of the marker expression density.

If several profiles are present in the CELL, CLUSTER, GATE objects, all profiles or combination of profiles will be plotted. If several comparison results are present in the RES objects, all comparison results will be plotted.

Comparison results contained in RES object can also be ploted using this function. In such representation, each bar corresponds to a marker with a height proportional to the marker distance or inclusion assessment, and where the bars are colored if they model a success.
}

