% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RES-class}
\alias{RES}
\alias{RES-class}
\title{RES class definition}
\description{
RES is a S4 object containing one or several comparison results.
}
\details{
This object mainly stores for each comparison result: the similarity or inclusion measure, the associated similarity or inclusion p-value, the markers similarity or inclusion measures and the marker similarity or inclusion successes (measures below a specific threshold).
}
\section{Slots}{

\describe{
\item{\code{comparisons}}{a data.frame containing for each comparison: the profile names, the similarity or inclusion measure, and the associated p-value}

\item{\code{comparisons.nb}}{is an integer indicating the number of comparisons}

\item{\code{markers}}{a character vector containing the marker names used in the comparisons}

\item{\code{marker.measures}}{a data.frame containing the marker similarity or inclusion measures for each comparison}

\item{\code{marker.successes}}{a data.frame containing the marker successes for each comparison}
}}

