% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_display.R
\name{res.dendro}
\alias{res.dendro}
\title{Dendrogram representation of the comparisons results}
\usage{
res.dendro(res, method = "complete")
}
\arguments{
\item{res}{a RES object}

\item{method}{a character of the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}
}
\value{
a ggplot plot object represnetating a dendrogram
}
\description{
CytoCompare can also represent these phenotype distances between the cell clusters using dendrograms. In such dendrogram, each leaf corresponds to a cell cluster, and the branching diagram represents the relationships of similarity among the cell clusters.
}
\details{
Hierarchical clustering can be constructed based on different linkage methods, such as the complete linkage represented here.
}
