% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_display.R
\name{res.graph}
\alias{res.graph}
\title{Circular graph representation of a RES object}
\usage{
res.graph(res, filename = "res.html", svgsize = 1000, pvalue.th = 0.05,
  color_edges = TRUE)
}
\arguments{
\item{res}{a RES object}

\item{filename}{a character specifying a file location where to save the HTML file of the representation}

\item{svgsize}{a numeric value specifying the size of the SVG representation in pixels}

\item{pvalue.th}{a numeric value specifying a p-value cutoff. Only the associations below this specific value will be returned}

\item{color_edges}{a boolean value specifying if edges but be colored}
}
\value{
none
}
\description{
Creates a circular graph representation of the comparison results. In such graph representation, each cytometry profile is represented by a node and links between the nodes represent significant similarities or inclusions between the profiles. Nodes are positioned on a circular layout and organized based on their object names.
}
\details{
The representation is provided as an interactive HTML file via a Scalable Vector Graphics (SVG) element created with the D3.js library. Users can interact with the representation by modifying the link tensions and the p-value cutoff for the significant similarities or inclusions.
}
