% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{set}
\alias{set}
\alias{[<-,MWEIGHTS,numeric,missing,numeric-method}
\alias{[<-,MWEIGHTS,character,missing,numeric-method}
\alias{[<-,MWEIGHTS,logical,missing,numeric-method}
\title{Change marker weights in a MWEIGHTS object}
\usage{
\S4method{[}{MWEIGHTS,numeric,missing,numeric}(x, i) <- value

\S4method{[}{MWEIGHTS,character,missing,numeric}(x, i) <- value

\S4method{[}{MWEIGHTS,logical,missing,numeric}(x, i) <- value
}
\arguments{
\item{x}{a MWEIGHTS object}

\item{i}{a numeric, logical or character vector}

\item{value}{a numeric vector containing the new marker weight values}
}
\value{
a S4 object of class MWEIGHTS
}
\description{
Sets the weights of the markers in a MWEIGHTS object.
}
\details{
Marker weights can be set based on their indexes or names in the MWEIGHTS object.
}
