% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prints.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,CLUSTER-method}
\alias{show,DENSITY-method}
\alias{show,MWEIGHTS-method}
\alias{show,RES-method}
\title{Textual preview for all S4 CytoCompare objects}
\usage{
\S4method{show}{CLUSTER}(object)

\S4method{show}{RES}(object)

\S4method{show}{MWEIGHTS}(object)

\S4method{show}{DENSITY}(object)
}
\arguments{
\item{object}{a CELL, CLUSTER, GATE, DENSITY, MWEIGHTS or RES object}
}
\value{
none
}
\description{
Shows a preview for a CELL, CLUSTER, GATE, DENSITY, MWEIGHTS or RES object.
}

