% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CELL-class}
\alias{CELL}
\alias{CELL-class}
\title{CELL class definition}
\description{
CELL is a S4 object containing one or several cell profiles.
}
\details{
This object mainly stores for each cell profile, the intensities of each marker.

The slot 'trans.para' is a named list contains different parameters depending of the transformation applied on the marker expression intensities. The scale (cofactor) of the arcsinh transformation function is parametrized using the 'arcsinh.scale' value. The shift of the log transformation function is parametrized using the 'log.shift' value and the base of the log transformation function is parametrized using the 'log.base' value. If no transformation function have been applied, the 'trans.para' slot is set to NULL.
}
\section{Slots}{

\describe{
\item{\code{name}}{a character indicating the internal name of the CELL object}

\item{\code{profiles}}{a character vector containing the names of the cell profiles}

\item{\code{profiles.nb}}{an integer value indicating the number of cell profiles}

\item{\code{markers}}{a character vector containing the marker names}

\item{\code{markers.nb}}{an integer value indicating the number of markers}

\item{\code{intensities}}{a numeric matrix containing the intensities of each marker for each cell profile}

\item{\code{trans}}{a character specifying the name of a transformation function applied on the marker expression intensities. Possible values are "arcsinh" for arc sin hyperbolic transformation, "log" for logarithmic transformation, or "none" for no transformation}

\item{\code{trans.para}}{a named list containing parameters of the transformation. Please refer to the details section for more details}

\item{\code{trans.exclude}}{a character vector containing the marker names for which no transformation has been applied on}

\item{\code{overview.function}}{a character specifying the name of a function to call when plotting the CELL object overview (please refer to the documentation of the 'plot()' function)}

\item{\code{layout}}{a numeric matrix that can be used to store the positions of cells in a 2-dimensional space (e.g. tSNE1 and tSNE2 dimensions provided by viSNE)}
}}

