% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export}
\alias{export}
\alias{export,CELL-method}
\alias{export,GATE-method}
\title{Exportation of CELL or GATE objects}
\usage{
export(object, filename)

\S4method{export}{CELL}(object, filename = "cells.FCS")

\S4method{export}{GATE}(object, filename = "gates.xml")
}
\arguments{
\item{object}{a CELL or GATE object}

\item{filename}{a character indicating the location of the FCS or GatingML-XML file to save}
}
\value{
none
}
\description{
Exports a CELL object into a FCS file or export a GATE object into a GatingML-XML file.
}

