% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.CLUSTER}
\alias{import.CLUSTER}
\title{Importation of cell cluster profiles from a tab separated file}
\usage{
import.CLUSTER(file, exclude = NULL)
}
\arguments{
\item{file}{a character specifying the location of a tab separated file to import}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}
}
\value{
a S4 object of class CLUSTER
}
\description{
Imports one or several cell cluster profiles from a tab separated file into a CLUSTER object. In this case, the marker expressions of each cluster are assumed to be normally distributed.
}
\details{
Tab separated file to import must contain for each cell cluster profile the means and the standard deviations of the expression markers and must be formatted as the following:\cr
* each row must represent a cell cluster profile;\cr
* each column must represent a marker;\cr
* each cell in the table must contain the marker expression means and the standard deviations for a given cell cluster separated by a semicolon;\cr
The first column must contain the cell cluster names and the first row must contain the marker names.

It is to note that `CLUSTER` objects constructed via the `import.CLUSTER()` function do not contain the densities of expression markers (please refer to the documentation of the `compare()` function).
}
