% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imports.R
\name{import.SPADE}
\alias{import.SPADE}
\title{Importation of cell cluster profiles from SPADE results}
\usage{
import.SPADE(path, exclude = NULL, trans = "arcsinh",
  trans.para = switch(trans, arcsinh = list(arcsinh.scale = 5), log =
  list(log.shift = "auto", log.base = 10), none = NULL), trans.exclude = NULL,
  bin.width = 0.05, extract.folder = NULL, extract.folder.del = FALSE,
  zip = FALSE, rescale = FALSE, rescale.quantiles = c(0, 1))
}
\arguments{
\item{path}{a character indicating the location to a zip or a folder containing the SPADE results}

\item{exclude}{a character vector containing the marker names to be excluded in the import procedure}

\item{trans}{a character specifying the name of a transformation function to apply on the marker expression intensities. Possible functions are "arcsinh" for arc sin hyperbolic transformation (default), "log" for logarithmic transformation, or "none" for no transformation}

\item{trans.para}{a named list containing parameters for the transformation. Please refer to the details section for more details}

\item{trans.exclude}{a character vector containing the marker names for which no transformation must be applied on}

\item{bin.width}{a numeric value indicating the width of the bins for the marker expression densities computations}

\item{extract.folder}{a folder path for extracting the SPADE zip archive (temporary folder by default)}

\item{extract.folder.del}{a logical value indicating if the extracted SPADE results should be removed after the extraction}

\item{zip}{a logical value that specify if the path specifies a zip file}

\item{rescale}{a logical specifying if marker expression intensities must be rescale between 0 and 1}

\item{rescale.quantiles}{a numeric vector of two values specifying the quantiles of the marker expression intensities used to rescale}
}
\value{
a S4 object of class CLUSTER
}
\description{
Imports one or several cell cluster profiles identified by the SPADE algorithm into a CLUSTER object.
}
\details{
SPADE is a popular visualization and analysis algorithm that identifies clusters of cells having similar expression profiles for selected markers using an agglomerative hierarchical clustering-based algorithm combined with a density-based down-sampling procedure (PMID:21964415). Given a set of FCS files (usually one file per sample), SPADE identifies cell clusters based on the whole dataset and provides then for each sample the amount of cells present within each cluster.

The 'rescale' parameter can be used to rescale the marker expression intensities from 0 to 1 (with respect to the distribution proportion). The rescaling can be performed based on minimal and maximal expression values or based on specified quantiles using the 'rescale.quantiles' parameter. This strategy is especialy usefull when comparing cell or cell cluster profiles obtained from different experimental/staining conditions.
}
