% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{methods}
\name{intersect}
\alias{intersect}
\alias{intersect,CLUSTER,CLUSTER-method}
\title{Identification of common markers between two cytometry objects}
\usage{
\S4method{intersect}{CLUSTER,CLUSTER}(x, y)
}
\arguments{
\item{x}{a CELL, CLUSTER or GATE object}

\item{y}{a CELL, CLUSTER or GATE object}
}
\value{
a S4 object of class MWEIGHTS
}
\description{
Identifies the common markers between two cytometry objects (CELL, CLUSTER or GATE objects) and store the results in a MWEIGHTS object.
}
\details{
The weight of each marker is set to 1 but can be changed afterwards using the function set.
}

