% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{RES-class}
\alias{RES-class}
\alias{RES}
\title{RES class definition}
\description{
RES is a S4 object containing one or several comparison results.
}
\details{
This object mainly stores for each comparison result: the associated distances (with the associated distance threshold) or inclusion p-value, and the marker successes. In the case of similarity comparisons, the aggregated distance between the two profiles and the marker distances are also stored in this object.
}
\section{Slots}{

\describe{
\item{\code{comparisons}}{a data.frame containing for each comparison: the profile names, the type of the comparison (similarity or inclusion), the aggregated distance (or NA in case of inclusion), the distance threshold used (\code{$D_th$}) and the associated p-value}

\item{\code{comparisons.nb}}{is an integer indicating the number of comparisons}

\item{\code{markers}}{a character vector containing the marker names used in the comparisons}

\item{\code{marker.weights}}{a numeric vector containing the weigths associated to each marker involded in the comparisons}

\item{\code{marker.distances}}{a data.frame containing the marker distances for each comparison (or NA in case of inclusion assessments)}

\item{\code{marker.successes}}{a data.frame containing the marker successes for each comparison}
}}

